doctor/requirements.txt
fastapi>=0.111.0
uvicorn[standard]>=0.29.0
requests>=2.31.0
psutil>=5.9.0
pyyaml>=6.0.1
